Use size_t for name_len member.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@3451 06827809-a52a-0410-b366-d66718629ded

+++ apps/snmpbulkget.c
@@ -84,7 +84,7 @@ int max_repetitions = 100;
 int non_repeaters = 0;
 struct nameStruct {
   oid name[MAX_OID_LEN];
-  u_long name_len;
+  size_t name_len;
 } *name, *namep;
 int names;
